requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cloudinary==1.33.0
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2.1
django-allauth==0.54.0
django-debug-toolbar==4.1.0
django-environ==0.10.0
django-taggit==4.0.0
Faker==18.9.0
idna==3.4
oauthlib==3.2.2
Pillow==9.5.0
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.16
